Refactor setting up date pickers
Previously, all code to set up the date pickers in
a text field was in the VC. Refactor some of the code
into a Date Picker extension for easy reusability.
Also tried to find a way to refactor objc functions into date
extension but believe that those only work in the same
file.